Switch to f64

+++ src/aabb.rs
@@ -19,7 +19,7 @@ impl AABB {
         &self.max
     }
 
-    pub fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> bool {
+    pub fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> bool {
         let mut min_t = t_min;
         let mut max_t = t_max;
